BinToAscii created and finished
