+++ Real/BinToASCIIConv.cpp
@@ -1,0 +1,39 @@
+#include <iostream>
+#include <string>
+#include <sstream>
+#include <bitset>
+using namespace std;
+
+class B2AConverter {
+    public:
+    string bits;
+
+    B2AConverter::B2AConverter(string bitstring){
+        string hold;
+        for (int i=0; i<bitstring.length(); i++){
+            if (bitstring[i]!=','){
+                hold += bitstring[i];
+            }
+        }
+        this->bits = hold;
+        convert();
+    }
+
+    string convert(){
+        stringstream sstream(this->bits);
+        string output;
+        while(sstream.good()){
+            bitset<8> bits;
+            sstream >> bits;
+            char c = char(bits.to_ulong());
+            output += c;
+            output +=',';
+        }
+    output = output.substr(0,output.length()-3); //c++ is bad so it includes the damn 11111111 char at the end for some god unknown reason
+    return output;
+    }
+};
+
+int main(){
+    B2AConverter c("00110001,00110010,00110011,00110100,00110101"); // This is the binary for 1,2,3,4,5 generated by FileToBinConv
+}
